Move FieldID to a separate module to fix the unrodered field destruction
