+++ legate/core/_legion/__init__.py
@@ -15,6 +15,7 @@
 from __future__ import annotations
 
 from .env import LEGATE_MAX_DIM, LEGATE_MAX_FIELDS
+from .field import FieldID
 from .future import Future, FutureMap
 from .geometry import Point, Rect, Domain
 from .operation import (
@@ -49,7 +50,6 @@ from .util import (
     dispatch,
     BufferBuilder,
     ExternalResources,
-    FieldID,
     FieldListLike,
     legate_task_preamble,
     legate_task_postamble,
